Can define HTTP check details on create/check
Add the specific HTTP details[1] and pass them as parameters for creation
or update.

We must handle some special cases, like username/password, which must
be passed as "auth": "user:pass" and HTTP headers.

Also added restrictions for ShouldContain and ShouldNotContain, which
are mutually exclusive and only one can be defined.

[1] https://www.pingdom.com/resources/api#MethodGet+Detailed+Check+Information