Separate methods for PostParams/PutParams
As we added details to HTTP Check, we found out that
the PUT and POST methods behave differently.

In the case of HTTP it is not allowed to POST params with
empty strings. To simplify the management of these differences
and ease add more check types in the future, we
split the method PostParams and PutParams. If the logic
is the same or similar, one can delegate on the other.

In the case of HTTP check, PostParams() simply filters
out attributes with empty string values.